+++ library/oneview_logical_interconnect_group.py
@@ -185,7 +185,6 @@ class LogicalInterconnectGroupModule(OneViewModule):
 
     def __replace_name_by_uris(self):
         map_template = self.data.get('interconnectMapTemplate')
-
         if map_template:
             map_entry_templates = map_template.get('interconnectMapEntryTemplates')
             if map_entry_templates:
@@ -201,11 +200,17 @@ class LogicalInterconnectGroupModule(OneViewModule):
                 allUplinkSets = self.__get_all_uplink_sets()
                 for uplinkSet in self.data['uplinkSets']:
                     networkNames = uplinkSet.pop('networkNames', None)
+                    networkSetNames = uplinkSet.pop('networkSetNames', None)
                     if networkNames and not uplinkSet.get('networkUris'):
                         uplinkSet['networkUris'] = []
                     if networkNames:
                         networkUris = [self.__get_network_uri(x) for x in networkNames]
                         uplinkSet['networkUris'].extend(networkUris)
+                    if networkSetNames and not uplinkSet.get('networkSetUris'):
+                        uplinkSet['networkSetUris'] = []
+                    if networkSetNames:
+                        networkSetUris = [self.__get_network_set(x) for x in networkSetNames]
+                        uplinkSet['networkSetUris'].extend(networkSetUris)
                     allUplinkSets = self.__update_existing_uplink_set(allUplinkSets, uplinkSet)
                 self.data['uplinkSets'] = allUplinkSets
             else:
@@ -214,11 +219,17 @@ class LogicalInterconnectGroupModule(OneViewModule):
     def __update_network_uri(self):
         for i in range(len(self.data['uplinkSets'])):
             networkNames = self.data['uplinkSets'][i].pop('networkNames', None)
+            networkSetNames = self.data['uplinkSets'][i].pop('networkSetNames', None)
             if networkNames and not self.data['uplinkSets'][i].get('networkUris'):
                 self.data['uplinkSets'][i]['networkUris'] = []
             if networkNames:
                 networkUris = [self.__get_network_uri(x) for x in networkNames]
                 self.data['uplinkSets'][i]['networkUris'].extend(networkUris)
+            if networkSetNames and not self.data['uplinkSets'][i].get('networkSetUris'):
+                self.data['uplinkSets'][i]['networkSetUris'] = []
+            if networkSetNames:
+                networkSetUris = [self.__get_network_set(x) for x in networkSetNames]
+                self.data['uplinkSets'][i]['networkSetUris'].extend(networkSetUris)
 
     def __update_existing_uplink_set(self, allUplinkSets, newUplinkSet):
         temp = True
@@ -248,6 +259,12 @@ class LogicalInterconnectGroupModule(OneViewModule):
         network_name = self.oneview_client.ethernet_networks.get_by('name', name)
         if network_name:
             return network_name[0]['uri']
+        return False
+
+    def __get_network_set(self, name):
+        network_set = self.oneview_client.network_sets.get_by('name', name)
+        if network_set:
+            return network_set[0]['uri']
         return False
 
     def __get_interconnect_type_by_name(self, name):